pyproject.toml
[tool.poetry]
name = "edged-emoji-maker"
version = "0.1.0"
description = "Edge your slack emojis with white (for dark mode)"
authors = ["dora <me@dora1998.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.18.1"
Pillow = "^7.0.0"
python-dotenv = "^0.10.3"
requests = "^2.22.0"
tqdm = "^4.41.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
